TRIANGLE SIDES
The mathematics teacher has recently taught the formula for computing the distance between 2 points to her students. She gave them an assignment on calculating the distance between the points. In order to prevent copying, she gave each student a different set of questions. But she found it very difficult to correct all 100 assignments. She seeks your help. Can you please help her out?
Given the vertices of a triangle ABC, determine the sides of the triangle AB, BC and AC. Can you write a C program to calculate the sides?
Input Format
Input consists of 6 integers. The first 2 integers correspond to the x-coordinate and y-coordinate of vertex A. The next 2 integers correspond to the x-coordinate and y-coordinate of vertex B. The next 2 integers correspond to the x-coordinate and y-coordinate of vertex C.
Output Format:
Refer Sample Input and Output for exact formatting specifications.
[All floating point values are displayed correct to 1 decimal place]
Sample Input and Output:
Enter the x-coordinate of vertex A
5
Enter the y-coordinate of vertex A
10
Enter the x-coordinate of vertex B
10
Enter the y-coordinate of vertex B
10
Enter the x-coordinate of vertex C
8
Enter the y-coordinate of vertex C
4
Length of side AB is 5.0
Length of side BC is 6.3
Length of side AC is 6.7