INTERCEPTS
Emily is a very popular Maths Teacher in School. She retires from her service today. Her 7th class students like her very much and they wanted to give her a grand farewell at school. The school Head Mistress is a very strict person and she didn't give permission for them to conduct the farewell in the school premises. The students decided to conduct the farewell at Emily Mam's house itself.
They know the street in which Emily mam lives. The student leader asked Emily Mam to tell her house number. Emily Mam's last class was on Equation for a straight line. She said that a straight line can be represented by the equation y=mx+c and you know how to find the x-intercept and y-intercept of the line and my house number is the sum of the x-intercept and y-intercept of the line.
The students were puzzled. Can you help the students find the house number of Emily Mam?
Given the values of m and c of the line equation y=mx+c, write a C program to find the sum of x-intercept and y-intercept.
Input Format:
Input consists of 2 integers. The first integer corresponds to m and the second integer corresponds to c.
Output Format:
Refer Sample Input and Output for exact formatting specifications.
[Assume that the inputs are such that the intercept values are always integers] 
Sample Input and Output:
Enter the value of m
5
Enter the value of c
10
The line equation is y=5x+10
The x intercept is -2
The y intercept is 10
The house number is 8